gally-agent/requirements.txt
crewai
crewai[tools]
fastapi
uvicorn
python-dotenv
google-generativeai
sqlalchemy
psycopg2-binary
gunicorn
